Add files via upload
Added the kfold estimator that Gloria added, and changed how the graphs look. This uses a for-loop which refits the same data 18 times, which I'll fix up tomorrow. That said, the new graphs it's making look interesting.  Also planning on writing the mean/CI generator as a function which just takes FIPS code as an input, which will let us iterate more easily.